requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
openpyxl>=3.1.2
pillow>=10.0.0
easyocr>=1.7.1
python-dateutil>=2.9.0.post0